add dag plus refactoring

+++ clickhouse/migrations/01_create_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS public_reviews
 (
     event_date      Date,
     language        String,
-    reviews_count   Int64,
+    reviews_count   Int32,
     min_score       Float32,
     avg_score       Float32,
     max_score       Float32,
